// src/scripts/enums.ts
export enum selectors {
  topBarSpace = ".top-bar-space",
}

export enum spritesheets {
  babyPurpleCow = "images/baby-purple-cow-spritesheet.png",
}

export enum AnimationTypes {
  Blink = "blink",
  Breathe = "breathe",
  Walk = "walk",
  Jump = "jump",
  Stand = "stand",
  Sleep = "sleep",
  Smell = "smell",
  Eat = "eat",
  Love = "love",
}